refactor: further refactor
